Placed handle function in Handler comp
