+++ components/organism/AdminTabs/AllCategories.jsx
@@ -1,31 +1,23 @@
 import { useContext } from 'react';
 import { Container } from 'react-bootstrap';
 import { GetRoute } from '../../../API/index';
-import { AlertContext } from '../../../Context/AlertContext';
 
 import { spinner } from '../../atom/spinner';
-import { card } from '../../molecule/card';
 import styles from './AllCategories.module.css';
 import { alert } from '../../atom/alert';
 import Api from '../../../API';
 
+import Handler from '../Handler';
+import { updateDeleteCategory } from '../../molecule/updateDeleteCategory';
+
+import { AdminContext } from '../../../Context/AdminContext';
+import { AlertContext } from '../../../Context/AlertContext';
+
 const AllCategories = () => {
+    const { handleConfirm, handleUpdate } = Handler();
+
     const alertContext = useContext(AlertContext);
-
     const fetcher = GetRoute('/api/category').data;
-    const { deleteRoute } = Api();
-
-    const handleUpdate = (id, title, image) =>
-        console.log('update was clicked', id, title, image);
-
-    const handleConfirm = (body) => {
-        let answer = window.confirm('Are you sure you want to delete');
-        if (answer) {
-            handleDelete(body);
-        }
-    };
-
-    const handleDelete = (body) => deleteRoute('/api/category', body);
 
     return (
         <Container>
@@ -35,15 +27,11 @@ const AllCategories = () => {
             {fetcher !== undefined
                 ? Object.entries(fetcher).map(([key, value]) => (
                       <div key={key} className={styles.flex}>
-                          <div
-                              style={{
-                                  width: '20rem',
-                                  height: '12rem',
-                                  padding: '2%',
-                              }}
-                          >
-                              {card(value, handleConfirm, handleUpdate)}
-                          </div>
+                          {updateDeleteCategory(
+                              value,
+                              handleConfirm,
+                              handleUpdate
+                          )}
                       </div>
                   ))
                 : spinner()}
